feat: fixed and replace all link from old to new components
