+++ src/pages/new.tsx
@@ -1,51 +1,172 @@
-import dbConnect from '../utils/dbConnect';
-import newFeedGenerator from '../utils/getNew';
-import { IVideo } from "src/models/Video";
-import React, { useState } from 'react';
-import { VideoCard } from 'src/components/VideoCard';
-import { Grid, Row } from 'src/components/Grid';
-import InfiniteScroll from "react-infinite-scroll-component";
+import VideosTitle from "@/components/VideosTitle";
+import Name from "@/components/user/Name";
+import { Box, Flex, Grid, GridItem, Image, Text } from "@chakra-ui/react";
+import React, { useState } from "react";
 
-export async function getServerSideProps() {
-  await dbConnect();
-  let created: IVideo[] = await newFeedGenerator({ languages: ['en'] })//languages);
-  return {
-    props: { created }
-  }
-}
-
-export default function New({ created }: { created: (IVideo & { payout: number; })[] }) {
-  const [createdVideos, setVideos] = useState(created);
-  const [page, setPage] = useState(1)
-
-  const getMoreVideos = async () => {
-    const lastVideo = createdVideos[createdVideos.length - 1]
-    fetch(`/api/new?` + new URLSearchParams({
-      page: `${page}`,
-      lastVideo: `${lastVideo.owner}/${lastVideo.permlink}`
-    }))
-      .then((res) => res.json())
-      .then((videos: (IVideo & { payout: number; })[]) => {
-        setPage(page + 1)
-        setVideos([...createdVideos, ...videos]);
-      })
-  };
-
+const New2 = () => {
+  const [videos, setVideos] = useState([
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfKKVdjVeExQngEq7iRUrnhc6nn53bkCVBsgFpHS7MyjhEFTNWBf73wdgVAiYs1bRwEXPsRfwGEUec54?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfKKVdjVbG93Si1Pof3Bp3SEoYnPNKtbHG2hmqSDhp9WVT7QEboptYxvDofQRog5ZtF2Mj6QXAxsbNPp?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfKKVdjVbFfGR7ZUTupbsAmkiGxmsGFMazc7fs8C9Da6r49Va7sXUUxwEpMs9Lp8kLufQoT7nHCojSg6?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfKKVdjVb4dHDxLG1Fefpe4k4AmqEgewJw6TMbjKkXmd6EpNKDDSY7jajEbHFUDkuNFFGj3QHTXGgwi6?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfKKVdjVeUe5z9NZYeLaAjAAD5cJkxV4PRBZnpDR1KknD4EJbrmG3Q9yNQgcEncQJwMeScuZQCF964RC?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfKKVdjVigXNQ9i2Sj17Sn3n278qQDgwZjFj63SVyowrUwLLy3b4XjctSSXPrUQzfGWnYFenbacvxXNn?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfHesJoVwVwkncjGuS3rGBYn5aS66yk9ripBpXdrZC3PDze6nrQmCgcnqLVtF7dmTz85o6rdCJk8XgTG?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfHesJoVwVwkncjGuS3rGBYn5aS66yk9ripBpXdrZC3PDze6nrQmCgcnqLVtF7dmTz85o6rdCJk8XgTG?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfHesJoVwVwkncjGuS3rGBYn5aS66yk9ripBpXdrZC3PDze6nrQmCgcnqLVtF7dmTz85o6rdCJk8XgTG?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfHesJoVwVwkncjGuS3rGBYn5aS66yk9ripBpXdrZC3PDze6nrQmCgcnqLVtF7dmTz85o6rdCJk8XgTG?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfHesJoVwVwkncjGuS3rGBYn5aS66yk9ripBpXdrZC3PDze6nrQmCgcnqLVtF7dmTz85o6rdCJk8XgTG?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfHesJoVwVwkncjGuS3rGBYn5aS66yk9ripBpXdrZC3PDze6nrQmCgcnqLVtF7dmTz85o6rdCJk8XgTG?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfHesJoVwVwkncjGuS3rGBYn5aS66yk9ripBpXdrZC3PDze6nrQmCgcnqLVtF7dmTz85o6rdCJk8XgTG?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfHesJoVwVwkncjGuS3rGBYn5aS66yk9ripBpXdrZC3PDze6nrQmCgcnqLVtF7dmTz85o6rdCJk8XgTG?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfHesJoVwVwkncjGuS3rGBYn5aS66yk9ripBpXdrZC3PDze6nrQmCgcnqLVtF7dmTz85o6rdCJk8XgTG?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfHesJoVwVwkncjGuS3rGBYn5aS66yk9ripBpXdrZC3PDze6nrQmCgcnqLVtF7dmTz85o6rdCJk8XgTG?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfHesJoVwVwkncjGuS3rGBYn5aS66yk9ripBpXdrZC3PDze6nrQmCgcnqLVtF7dmTz85o6rdCJk8XgTG?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfHesJoVwVwkncjGuS3rGBYn5aS66yk9ripBpXdrZC3PDze6nrQmCgcnqLVtF7dmTz85o6rdCJk8XgTG?format=jpeg&mode=cover&width=340&height=191",
+    },
+  ]);
   return (
-    <Grid>
-      <h1>New</h1>
-      <InfiniteScroll
-        dataLength={createdVideos.length}
-        next={getMoreVideos}
-        hasMore={true}
-        loader={<h3>Loading...</h3>}
-        endMessage={<h4>Nothing more to show</h4>}
-      ><Row>
-        {createdVideos.map((video: IVideo & { payout: number; }) => (
-          <VideoCard key={`${video.owner}/${video.permlink}`} {...video} />
+    <Box>
+      <Box backgroundColor={"#E8E8E8"} padding="20px">
+        <Text as="h1" fontWeight={"300 !important"}>
+          NEW VIDEOS
+        </Text>
+      </Box>
+      <Grid padding={"20px"} templateColumns={{
+          base: "repeat(2, 1fr)",
+          md: "repeat(3, 1fr)",
+          lg: "repeat(3, 1fr)",
+          xl: "repeat(6, 1fr)",
+        }} gap={6}>
+        {/* {videos.length} */}
+        {videos.map((video, index) => (
+          <GridItem w="100%" h="100%" key={index}>
+            <Image
+              padding={"5px"}
+              backgroundColor={"#222 !important"}
+              alt="test"
+              src={`${video.thumbnail}`}
+            />
+            <VideosTitle title={`${video.title}`} />
+            <Name username={`${video.username}`} />
+            <Text as="p" margin={"1px"}>
+              a day ago
+            </Text>
+            <Text fontWeight={"bold"} as="p">
+              $ 10.10
+            </Text>
+          </GridItem>
         ))}
-        </Row>
-      </InfiniteScroll>
-    </Grid>
-  )
-}
+      </Grid>
+    </Box>
+  );
+};
+
+export default New2;
