+++ src/pages/trends.tsx
@@ -1,49 +1,107 @@
-import dbConnect from '../utils/dbConnect';
-import trendingFeedGenerator from '../utils/getTrending';
-import { IVideo } from "../models/Video";
-import React, { useState } from 'react';
-import { VideoCard } from '../components/VideoCard';
-import { Grid, Row } from '../components/Grid';
-import InfiniteScroll from "react-infinite-scroll-component";
+import VideosTitle from "@/components/VideosTitle";
+import Name from "@/components/user/Name";
+import { Box, Flex, Grid, GridItem, Image, Text } from "@chakra-ui/react";
+import { css } from "@emotion/react";
+import React, { useState } from "react";
 
-export async function getServerSideProps() {
-  await dbConnect();
-  let trending = await trendingFeedGenerator({ languages: ['en'] })//languages);
-  return {
-    props: { trending }
-  }
-}
-
-export default function Trending({ trending }: { trending: (IVideo & { payout: number; })[] }) {
-  const [trendingPosts, setPosts] = useState(trending);
-  const [page, setPage] = useState(1)
-
-  const getMoreVideos = async () => {
-    fetch(`/api/trending?` + new URLSearchParams({
-      page: `${page}`
-    }))
-      .then((res) => res.json())
-      .then((videos: (IVideo & { payout: number; })[]) => {
-        setPage(page + 1)
-        setPosts([...trendingPosts, ...videos]);
-      })
-  };
-
+const NewComers2 = () => {
+  const [videos, setVideos] = useState([
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfHesJoVwVwkncjGuS3rGBYn5aS66yk9ripBpXdrZC3PDze6nrQmCgcnqLVtF7dmTz85o6rdCJk8XgTG?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfKKVdjVfJRLvVE5BvTWSaKDuG2V5WY8voUDnYRYyQCQ2mP5vrsECF8CejKSNS4BouWcu7EZSrWtZ2EA?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfKKVdjViUUPgY24GUYeConGrus7C4UPGS2f5LkMjMShfeap5qDU7Guy2DR83RwGUsHZrU28PNmuwEuc?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfKKVdjVb4dHDxLG1Fefpe4k4AmqEgewJw6TMbjKkXmd6EpNKDDSY7jajEbHFUDkuNFFGj3QHTXGgwi6?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfKKVdjVeUe5z9NZYeLaAjAAD5cJkxV4PRBZnpDR1KknD4EJbrmG3Q9yNQgcEncQJwMeScuZQCF964RC?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfKKVdjVigXNQ9i2Sj17Sn3n278qQDgwZjFj63SVyowrUwLLy3b4XjctSSXPrUQzfGWnYFenbacvxXNn?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfHesJoVwVwkncjGuS3rGBYn5aS66yk9ripBpXdrZC3PDze6nrQmCgcnqLVtF7dmTz85o6rdCJk8XgTG?format=jpeg&mode=cover&width=340&height=191",
+    },
+    {
+      title:
+        "The new update of LIKETU 📱 (subtitled) - Conoce la nueva actualización de LIKETU",
+      username: " hiveredcarpet",
+      thumbnail:
+        "https://images.hive.blog/p/99pyU5Ga1kwr5bsMXthzYLbcngN4W2P8NtU9TWTdHC3HaQbjuuRfHesJoVwVwkncjGuS3rGBYn5aS66yk9ripBpXdrZC3PDze6nrQmCgcnqLVtF7dmTz85o6rdCJk8XgTG?format=jpeg&mode=cover&width=340&height=191",
+    },
+  ]);
   return (
-    <Grid>
-      <h1>Trending</h1>
-      <InfiniteScroll
-        dataLength={trendingPosts.length}
-        next={getMoreVideos}
-        hasMore={true} // TODO: trending can end when 7 days of content has been accessed
-        loader={<h3>Loading...</h3>}
-        endMessage={<h4>Nothing more to show</h4>}
-      ><Row>
-        {trendingPosts.map((video: IVideo & { payout: number; }) => (
-          <VideoCard key={`${video.owner}/${video.permlink}`} {...video} />
+    <Box>
+      <Box backgroundColor={"#E8E8E8"} padding="20px" paddingBottom={"0px"}>
+        <Text as="h1" fontWeight={"200 !important"}>
+          TRENDING VIDEOS
+        </Text>
+      </Box>
+      <Grid
+        padding={"20px"}
+        templateColumns={{
+          base: "repeat(2, 1fr)",
+          md: "repeat(2, 1fr)",
+          lg: "repeat(3, 1fr)",
+          xl: "repeat(4, 1fr)",
+        }}
+        gap={4}
+      >
+        {/* {videos.length} */}
+        {videos.map((video, index) => (
+          <GridItem w="100%" h="100%" key={index}>
+            <Image
+              padding={"5px"}
+              backgroundColor={"#222 !important"}
+              alt="test"
+              src={`${video.thumbnail}`}
+            />
+            <VideosTitle title={`${video.title}`} />
+            <Name username={`${video.username}`} />
+            <Text as="p" margin={"1px"}>
+              a day ago
+            </Text>
+            <Text fontWeight={"bold"} as="p">
+              $ 10.10
+            </Text>
+          </GridItem>
         ))}
-        </Row>
-      </InfiniteScroll>
-    </Grid>
-  )
-}
+      </Grid>
+    </Box>
+  );
+};
+
+export default NewComers2;
